Separate data loading logic to `Source`s implementations.
Before this change, we were developing mostly for PostgreSQL until we had an idea of what we wanted. Now that postgres source works well we move specific logic back to the Source, enabling other source-specific logic.

One key point is that now the Source contains the pool, later we might have to revisit this since depending on the user preferences e.g. tls vs notls, we have to construct the pools differently and types will change.

We also added a global static PERF_LOGGER in perf_logger.rs, which is our old trusty perf_logger but reusable with a lock.